Revert "move authentication function to util"
This reverts commit 80c879e6fc47cb724a15ad471d38c14f9562cd13.
