+++ src/app.js
@@ -6,6 +6,7 @@ const commentRouter = require("./controllers/comment-routes.js");
 const generalService = require("./service/general-service.js");
 const { logger } = require("./util/logger.js"); 
 
+
 const PORT = 3000;
 
 app.use(express.json());
