+++ src/service/user-service.js
@@ -102,6 +102,14 @@ async function createProfile(item, uuid, creation_date) {
     }
 }
 
+async function decodeJWT(token){
+    try{
+        const user = await jwt.verify(token, secretKey)
+        return user;
+    }catch(err){
+        console.error(err);
+    }
+}
 
 module.exports = {
     register,
